requirements.txt
coverage
pytest
mysql-connector-python